# public/openapi.yaml
openapi: 3.0.3
info:
  title: HookStep / Talent Platform API
  description: REST API for the hiring platform — jobs, applications, profile, and company dashboard.
  version: 1.0.0
servers:
  - url: /
    description: Current origin (relative)
...
paths:
...
  /api/jobs/{id}:
    get:
      tags: [Jobs]
      summary: Get job by ID (public)
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Job details (safe fields only)
          content:
            application/json:
              schema: { $ref: '#/components/schemas/JobPublic' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '404': { $ref: '#/components/responses/NotFound' }
...
  /api/applications:
    get:
      tags: [Applications]
      summary: List my applications (candidate)
      security: [{ session: [] }]
      responses:
        '200':
          description: Applications for the authenticated candidate
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/ApplicationWithJob' }
        '401': { $ref: '#/components/responses/Unauthorized' }
    post:
      tags: [Applications]
      summary: Submit application (candidate only)
      security: [{ session: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [jobId, resumeUrl]
              properties:
                jobId: { type: string }
                resumeUrl: { type: string, format: uri }
                motivation: { type: string }
                currentCTC: { type: string }
                expectedCTC: { type: string }
                currentCurrency: { type: string }
                expectedCurrency: { type: string }
                noticePeriod: { type: string }
                city: { type: string }
      responses:
        '201':
          description: Application created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Application' }
        '400':
          description: Missing fields, invalid URL, already applied, or job not active
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
...
  /api/company/applications/{id}:
    get:
      tags: [Company]
      summary: Get application by ID (recruiter, own company only)
      security: [{ session: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Application with job and candidate
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ApplicationWithCandidate' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
    patch:
      tags: [Company]
      summary: Update application status (recruiter)
      security: [{ session: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [status]
              properties:
                status: { type: string, enum: [APPLIED, SHORTLISTED, REJECTED, INTERVIEW, HIRED] }
                interviewScheduledAt: { type: string, format: date-time }
      responses:
        '200':
          description: Updated application
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Application' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }

  # ----- Company dashboard -----
...
components:
  securitySchemes:
    session:
      type: apiKey
      in: cookie
      name: next-auth.session-token
      description: NextAuth session cookie (sign in via the app)
  responses:
    Unauthorized:
      description: Not authenticated
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
      headers: {}
    Forbidden:
      description: Authenticated but not allowed (e.g. wrong role)
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
  schemas:
    Error:
      type: object
      properties:
        error: { type: string }
    Job:
      type: object
      properties:
        id: { type: string }
        title: { type: string }
        company: { type: string }
        location: { type: string }
        description: { type: string }
        employmentType: { type: string }
        category: { type: string }
        skills: { type: array, items: { type: string } }
        salary: { type: string }
        currency: { type: string }
        status: { type: string }
        experienceMin: { type: integer }
        experienceMax: { type: integer }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    JobPublic:
      type: object
      description: Same as Job but only safe/public fields (no companyId)
      properties:
        id: { type: string }
        title: { type: string }
        company: { type: string }
        location: { type: string }
        description: { type: string }
        employmentType: { type: string }
        category: { type: string }
        skills: { type: array, items: { type: string } }
        salary: { type: string }
        currency: { type: string }
        status: { type: string }
        experienceMin: { type: integer }
        experienceMax: { type: integer }
        createdAt: { type: string, format: date-time }
    Application:
      type: object
      properties:
        id: { type: string }
        jobId: { type: string }
        candidateId: { type: string }
        companyId: { type: string }
        resumeUrl: { type: string }
        motivation: { type: string }
        status: { type: string }
        appliedAt: { type: string, format: date-time }
        interviewScheduledAt: { type: string, format: date-time, nullable: true }
...
    ApplicationWithJob:
      type: object
      description: Application with nested job (for candidate list)
      properties:
        id: { type: string }
        job: { type: object, properties: { id: { type: string }, title: { type: string }, company: { type: string }, experienceMin: {}, experienceMax: {} } }
        appliedAt: { type: string, format: date-time }
        status: { type: string }
    ApplicationWithCandidate:
      type: object
      description: Application with job and candidate (for recruiter view)
      properties:
        id: { type: string }
        job: { $ref: '#/components/schemas/Job' }
        candidate: { type: object, properties: { id: {}, name: {}, email: {}, profileImageUrl: {} } }
        resumeUrl: { type: string }
        motivation: { type: string }
        status: { type: string }
        appliedAt: { type: string, format: date-time }
        interviewScheduledAt: { type: string, format: date-time, nullable: true }
...